Build out core_arch for x86_64
This has a separate struct for each SIMD capability level. That will
make it easier to split off lower levels if needed, and also a good
basis for higher levels.
