+++ src/core_arch/mod.rs
@@ -5,3 +5,6 @@
 
 #[cfg(target_arch = "aarch64")]
 pub mod aarch64;
+
+#[cfg(target_arch = "x86_64")]
+pub mod x86_64;
